Don’t inherit from View to implement Component
This commit refactors out all shared functionality between Component and
View into separate mixins and the CoreView base class. This refactor
accomplishes a few different goals.

First, Views will be deprecated and removed, likely in the 2.0 release.
By removing the View dependency from Component, we can simply delete the
View file once the 2.0 branch occurs.

Second, new angle bracket components (`<my-component>`) should not carry
legacy semantics that incur a performance cost. By separating out the
component concerns into mixins, we can only apply those mixins to the
Component base class for applications that use the legacy
`{{my-component}}` invocation. (This commit does not make that change
but rather sets the stage for it.)
