+++ packages/ember-views/lib/streams/utils.js
@@ -3,7 +3,7 @@ import { get } from "ember-metal/property_get";
 import { isGlobal } from "ember-metal/path_cache";
 import { fmt } from "ember-runtime/system/string";
 import { read, isStream } from "ember-metal/streams/utils";
-import View from "ember-views/views/view";
+import CoreView from "ember-views/views/core_view";
 import ControllerMixin from "ember-runtime/mixins/controller";
 
 export function readViewFactory(object, container) {
@@ -22,7 +22,7 @@ export function readViewFactory(object, container) {
     viewClass = value;
   }
 
-  Ember.assert(fmt(value+" must be a subclass or an instance of Ember.View, not %@", [viewClass]), View.detect(viewClass) || View.detectInstance(viewClass));
+  Ember.assert(fmt(value+" must be a Component or View, not %@", [viewClass]), CoreView.detect(viewClass) || CoreView.detectInstance(viewClass));
 
   return viewClass;
 }
